+++ CHANGELOG.md
@@ -5,6 +5,7 @@ History of all released versions and changes
 ## v1.3.3
 ### Fixed
 - The DragViewportButton was not highlighted when selected
+- Fix a bug where the extension crashes when a container got selected without any exposed ports in the docker container connection-type.
 
 ## v1.3.2
 ### Fixed